v3.0.4
1. Changed baud rate configuration to allow direct input as an integer instead of selecting from an enum.
2. Added an automatic log generation feature. Logs are deleted and regenerated when they exceed 1MB in size.

+++ source/SerialPort/Ports/COMManager.cpp
@@ -12,7 +12,7 @@ COMManager::COMManager(LogManager& log, ConnectionConfig& config, DataManager& d
 
 	// COM 포트 개방용 데이터 저장
 	_portName     = _config.deviceName;
-	_baudRate     = (int)_config.baudRate;
+	_baudRate     = _config.baudRate;
 	_dataBits     = (char)_config.dataBit;
 	_parity       = (char)_config.parity;
 	_stopBits     = (char)_config.stopBit;
